added sidebar controller
