+++ src/main/resources/static/Views/templates/Sidebar.html
@@ -11,7 +11,7 @@
 
 		<nav id="sidebar">
 
-			<a href="@{/accueil}">
+			<a href="/accueil">
 				<div class="sidebar-header">
 					<h1 class="d-none d-md-block">Plan-T</h1>
 					<strong>PL.</strong>
@@ -19,21 +19,21 @@
 			</a>
 
 			<ul class="list-unstyled components">
-				<li class="active"><a href="@{/accueil}" class="page"> <img class="hidden " src="./Images/accueil.png">
+				<li class="active"><a href="/accueil" class="page"> <img class="hidden " src="./Images/accueil.png">
 						<div class="Titre d-none d-md-block">
 							<img src="./Images/accueil.png">&nbsp;&nbsp;&nbsp;&nbsp;Accueil
 						</div>
 					</a></li>
 
 
-				<li><a href="@{/calendrier}" class="page"> <img class="hidden" src="./Images/calendrier.png">
+				<li><a href="/calendrier" class="page"> <img class="hidden" src="./Images/calendrier.png">
 						<div class="Titre d-none d-md-block">
 							<img src="./Images/calendrier.png">&nbsp;&nbsp;&nbsp;&nbsp;Calendrier
 						</div>
 					</a></li>
 
 				<li>
-					<a href="@{/login}" class="page">
+					<a href="/jardin" class="page">
 						<img class="hidden" src="./Images/plan.png">
 						<div class="Titre d-none d-md-block">
 							<img src="./Images/plan.png">&nbsp;&nbsp;&nbsp;&nbsp;Jardin
@@ -41,21 +41,21 @@
 					</a>
 				</li>
 
-				<li><a href="@{/arrosage}" class="page"> <img class="hidden" src="./Images/arrosage.png">
+				<li><a href="/arrosage" class="page"> <img class="hidden" src="./Images/arrosage.png">
 						<div class="Titre d-none d-md-block">
 							<img src="./Images/arrosage.png">&nbsp;&nbsp;&nbsp;&nbsp;Arrosage
 						</div>
 					</a></li>
 
-				<li><a href="@{/formation}" class="page"> <img class="hidden" src="./Images/cours.png">
+				<li><a href="/formation" class="page"> <img class="hidden" src="./Images/cours.png">
 						<div class="Titre d-none d-md-block">
 							<img src="./Images/cours.png">&nbsp;&nbsp;&nbsp;&nbsp;Formation
 						</div>
 					</a></li>
 
-				<li><a href="@{/moncompte}" class="page"> <img class="hidden" src="./Images/compte.png">
+				<li><a href="/moncompte" class="page"> <img class="hidden" src="./Images/compte.png">
 						<div class="Titre d-none d-md-block">
-							<img src="./Images/compte.png">&nbsp;&nbsp;&nbsp;&nbsp;Connexion
+							<img src="./Images/compte.png">&nbsp;&nbsp;&nbsp;&nbsp;Mon compte
 						</div>
 					</a></li>
 
@@ -63,7 +63,7 @@
 
 			<div class="sidebar-footer">
 				<ul class="list-unstyled components">
-					<li><a href="@{/infos}"> <img id="image_infos" src="./Images/infos.png">
+					<li><a href="@infos"> <img id="image_infos" src="./Images/infos.png">
 						</a></li>
 				</ul>
 			</div>
